chore(docs): added docs

+++ solidity/library/TheLodgeConfig.sol
@@ -3,32 +3,38 @@ pragma solidity >=0.8.4 <0.9.0;
 
 import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 
+/// @notice Helper library to hold the configuration structs
+/// used by TheLodge contracts.
 library TheLodgeConfig {
+  /// @notice TheLodge configuration.
   struct Config {
     SaleConfig saleConfig;
     RevelationConfig revelationConfig;
     URIConfig uriConfig;
   }
 
+  /// @notice TheLodgeSaleHandler configuration.
   struct SaleConfig {
     string tokenName;
     string tokenSymbol;
     address oracle;
     uint32 maxDelay;
     uint256 nftPrice;
-    uint16 maxTokensPerAddress;
+    uint8 maxTokensPerAddress;
     IERC20 alternativePaymentToken;
     uint256 saleStartTimestamp;
     uint256 openSaleStartTimestamp;
     bytes32 merkleRoot;
   }
 
+  /// @notice TheLodgeRevelationHandler configuration.
   struct RevelationConfig {
     address vrfCoordinator;
     bytes32 keyHash;
     uint64 subId;
   }
 
+  /// @notice The URI configuraiton used by TheLodgeTokenInspectorHandler.
   struct URIConfig {
     string baseURI;
     string unrevealedURI;
